LU-5680 tests: Remove use of /etc/motd from sanity-hsm
In sanity-hsm tests, /etc/motd is copied to the file system
and archived. /etc/motd may not exist in some Linux distributions
and, if exists, may be of size 0 and may defeat the use of
bandwidth control for timing. Uses of /etc/motd are replaced
with temp file

Test-Parameters: trivial testlist=sanity-hsm

Lustre-change: https://review.whamcloud.com/14021
Lustre-commit: 715c7e9166a34c1f2b781f3314d8b7c02e10950c

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I2f287db6d422ce20bca037ca69e24179c7e48144
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Signed-off-by: Minh Diep <minh.diep@intel.com>
Reviewed-on: https://review.whamcloud.com/30795
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>