LU-10485 lustre: move LA_* flags to lustre_user.h
The LA_* flags are written to disk as part of the ChangeLog records
in mdd_attr_set_changelog(), which means they are now part of the
on-disk and network protocol, and cannot be changed (at least not the
first 12 bits that are written).
They need to be moved to lustre_user.h.

Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I9fc92e01301e70f0f4e5cd74135b9b2079d63658
Reviewed-on: https://review.whamcloud.com/30825
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>