LU-10461 tests: call exit in the skip routine
There are many reasons to not run, or skip, a test; the test
may require a certain number of servers or a certain Lustre version.
In these cases, the skip() or skip_env() routine is called. When we
call skip, the intention is to exit the routine early. Thus, call
‘exit 0’ at the end of the skip() routine.

Some calls to skip() are changed to skip_env() when a test is being
skipped due to the Lustre configuration or test environment.

Backported from b2_12 branch to b2_10 (partial backport: sanity.sh
changes are not included here):
Lustre-change: https://review.whamcloud.com/30964
Lustre-commit: 33aad7829de56353cca5bbd082d95d6821a7be9c

Signed-off-by: James Nunez <jnunez@whamcloud.com>
Change-Id: I42fd9535c0a803f334dfc5685f451a6bdc85e84b
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-on: https://review.whamcloud.com/34023
Tested-by: Jenkins
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>