LU-10210 tests: Add lustre_routes_conversion script in PATH
Fix the typo in test-framework.sh so that test_67 in
conf-sanity.sh find lustre_routes_conversion script
when running out of build tree.

Fixing the typo in test-framework.sh for exporting
LUSTRE_ROUTES_CONVERSION to pick the lustre_routes_conversion
script from $LUSTRE/scripts so that it is visible when running
out of build tree.

Change-Id: I1bd9a28e036b9c7b60eaa9886e641610d414c8ee
Test-Parameters: trivial testlist=conf-sanity
Signed-off-by: Sonia Sharma <sonia.sharma@intel.com>
Reviewed-on: https://review.whamcloud.com/31173
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: James Nunez <james.a.nunez@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>