LU-10850 fld: handle empty ranges in fld_index_init()
In fld_index_init(), ignore the empty range corresponding to the
zeroed-out phantom record that is included in newly created ldiskfs
IAM indexes. (This phantom record will be removed in a subseuent
patch.) Ensure that lfs_new will be set when the FLD index was empty
of valid ranges. (Previously it was properly set on ZFS backed targets
but not on ldiskfs.)

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Change-Id: I9850eac686e1488ffbd627f100075ce6f29200d8
Reviewed-on: https://review.whamcloud.com/31871
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Fan Yong <fan.yong@intel.com>
Reviewed-by: Mike Pershin <mike.pershin@intel.com>
Reviewed-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>