LU-10707 ksocklnd: revert back to jiffies
It was found for the case of routers which have LNet enabled
ethernet and infiniband that they couldn't communicate properly
with other. This was due to ko2iblnd still using jiffies and
the socklnd driver using time64_t. After some discussion it was
deceided that the best thing to do is roll back instead of
additional changes.

Revert "LU-9397 ksocklnd: move remaining time handling to 64 bits"

This reverts commit 59c25356208706b4dd0920bc7754f21e2db14a0f.

Change-Id: I42fd5620ab8131cd5fdf4bf0ef15553f6eabe550
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-on: https://review.whamcloud.com/32015
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>